add LM scaling property
LM scaling is intended to control the ratio between the LM strength and
the acoustic model strength *after* decoding, when the best path in the
lattice is generated.
Setting this to a value higher than 1.0 (default) makes the LM more
prominent in the lattice path weighting, so the best path may be more
aligned with the LM, and not with the acoustic model.

Signed-off-by: Amit Beka <amit.beka@gmail.com>