add nbest-results signal with configurable number of transcriptions
Add the 'nbest-results' signal which enables configurable number of
transcriptions to be returned.

The new signal emits a GObject named KaldiResult, which has a single
property named "texts", which is a newline-delimited string of all the
transcriptions. Other methods of passing variable number of results
(GArray, JSON, GObject with GArray) have failed until now because they
were opaque to PyGTK, or had memory issues, or for unknown reasons.
Passing a configurable GArray of results, or a GObject with dynamic
array of results is prefereable to the current approach, but I couldn't
make it work.

The main filter got a new property, 'num-nbest', which indicates the
number of transcriptions to output in the nbest-results signal. If it is
set to 0 (the default), the signal isn't called. Note that Kaldi can
emit less transcriptions than this number (based on the lattice itself).

Signed-off-by: Amit Beka <amit.beka@gmail.com>