add silence weighting to ivector extraction
for some speakers, ivector extraction can go wrong for multiple reasons.
in order to overcome the problem, a method for reducing the weight of
silence phones and long phonemes suspected as silence is introduced.

This option is available only in the unthreaded decoder - I haven't seen
an implementation for the threaded decoder yet.

Signed-off-by: Amit Beka <amit.beka@gmail.com>