prepare-models: Save model files under `models` directory.
Keeping all the downloaded files would make it easy to ignore those
files in the .gitignore instead of ignoring each type of file, e.g *.txt,
*.int, etc.
NOTE: The files under the `ivector_extractor` and `conf` directory are
not saved under the `models` as it would require to change the path in
those config files. These config files are not under source control and
are downloaded from an external source.